Fix wrong config.json location
The config.json was expected to be in the directory you are currently starting the rta.exe from, which will not always be the directory the with the rta.exe. This fix makes it so as long as the config.json is next to the rta.exe you can run from anywhere.